requirements.txt
zstandard
zstd
pillow
multiprocessing
opencv-python -- done
fpstimer
